organize test files to mirror src files
