+++ test/unit/plugins/parser/resourcelink.test.js
@@ -1,5 +1,5 @@
 const expect = require('expect');
-const parser = require('../../../src/cdoFlavoredParser');
+const parser = require('../../../../src/cdoFlavoredParser');
 
 describe('resourcelink', () => {
   describe('render', () => {
